+++ projects/ngx-opendesign/src/public-api.ts
@@ -92,6 +92,7 @@ export * from './lib/charts/bar-chart';
 export * from './lib/charts/pie-chart';
 export * from './lib/charts/donut-chart';
 export * from './lib/charts/stacked-bar';
+export * from './lib/charts/polar-chart';
 export * from './lib/charts/heatmap-chart';
 export * from './lib/charts/calendar-heatmap';
 export * from './lib/editor/code-highlight';